Supports running pre-configured commands with .smartest.*
If you need a specific command to be run, perhaps because you're using Docker
and cannot run tests from your Vim, define a file called `.smartest.#{type}`.

For example, for a Ruby project, including Rails projects, create the file
`.smartest.ruby` and write in it what you want Smartest to run. Ideally put it
in `.gitignore`.

For example, if you're using Docker and is running the bash in a Tmux pane (e.g
1), you could press `<leader>r` and have it run it on the pane 1. Define
`.smartest.ruby`:

    tmux send-keys -t 1 "time rspec --color --order random $filename" ENTER