Supports Java with Maven and -Dtest=*
When running a specific class it will now come up with the -Dtest parameter to
use in the runner. It does it by reading each line of the test file and finding
class names before the line the cursor is in. One the test is run, that class is
used.